+++ src/Http/Action/Posts/DeletePost.php
@@ -2,10 +2,12 @@
 
 namespace XenonCodes\PHP2\Http\Action\Posts;
 
+use XenonCodes\PHP2\Blog\Exceptions\AuthException;
 use XenonCodes\PHP2\Blog\Exceptions\PostNotFoundException;
 use XenonCodes\PHP2\Blog\Repositories\PostsRepository\PostsRepositoryInterface;
 use XenonCodes\PHP2\Blog\UUID;
 use XenonCodes\PHP2\Http\Action\ActionInterface;
+use XenonCodes\PHP2\Http\Auth\TokenAuthenticationInterface;
 use XenonCodes\PHP2\Http\ErrorResponse;
 use XenonCodes\PHP2\Http\Request;
 use XenonCodes\PHP2\Http\Response;
@@ -14,20 +16,29 @@ use XenonCodes\PHP2\Http\SuccessfulResponse;
 class DeletePost implements ActionInterface
 {
     public function __construct(
-        private PostsRepositoryInterface $postsRepository
-    )
-    {
+        private PostsRepositoryInterface $postsRepository,
+        private TokenAuthenticationInterface $authentication
+    ) {
     }
-
 
     public function handle(Request $request): Response
     {
         try {
-            $postUuid = $request->query('uuid');
-            $this->postsRepository->get(new UUID($postUuid));
+            $authenticatedUser = $this->authentication->user($request);
+        } catch (AuthException $exception) {
+            return  new  ErrorResponse($exception->getMessage());
+        }
 
+        try {
+            $postUuid = $request->query('uuid');
+            $post = $this->postsRepository->get(new UUID($postUuid));
         } catch (PostNotFoundException $error) {
             return new ErrorResponse($error->getMessage());
+        }
+
+        // Проверьте, является ли аутентифицированный пользователь автором сообщения
+        if ($authenticatedUser->getId() !== $post->getAuthor()->getId()) {
+            return new ErrorResponse('Вы не уполномочены удалять этот пост!');
         }
 
         $this->postsRepository->delete(new UUID($postUuid));
